Introduce frontend locking
Currently, in order to run dpkg, frontends have to release
the lock before invoking dpkg and re-acquire it afterwards,
leaving a short time where the database is unlocked and a
different dpkg process or frontend could lock it.

Frontend locking addresses the problem by creating a frontend-lock
lock file that is acquired by the frontend and not released for
dpkg invocations. Thus, multiple frontends cannot race for the
database lock.

This change extends the frontend lock to dpkg itself, acquiring
it whenever the variable DPKG_FRONTEND_LOCKED is not set, so that
a user manually running dpkg or a frontend not supporting this
protocol cannot interfere with a currently running frontend.