WIP: Initial integration with dpkg unpack
This might work, or not, I don't know, but it's a good
start to make the idea clearer. This unfortunately forks
two subprocesses, one for ddelta, and another copying from
the tarball to ddelta, as we would otherwise need to have a
loop that slices from tarball to ddelta and then from ddelta
to file.

The clean up is not really handled yet, and this also only
relies on the filename ending in .deltadeb to enable delta
support.